Pull *ComponentImplementationBuilder out of ComponentImplementationFactory, now that they are static classes.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=227751815
